+++ src/mesido/potential_errors.py
@@ -13,6 +13,7 @@ class MesidoAssetIssueType(Enum):
     COLD_DEMAND_POWER = "cold_demand.power"
     HEAT_DEMAND_TYPE = "heat_demand.type"
     ASSET_PROFILE_CAPABILITY = "asset_profile.capability"
+    HEAT_EXCHANGER_TEMPERATURES = "heat_exchanger.temperature"
 
 
 class PotentialErrors: